+++ src/Pim/Bundle/CatalogBundle/Form/Subscriber/FilterLocaleValueSubscriber.php
@@ -81,7 +81,7 @@ class FilterLocaleValueSubscriber implements EventSubscriberInterface
      */
     private function isTranslatable($attribute)
     {
-        return $attribute && $attribute->getTranslatable();
+        return $attribute && $attribute->isTranslatable();
     }
 
     /**