+++ src/Pim/Bundle/ImportExportBundle/Normalizer/ProductNormalizer.php
@@ -159,7 +159,7 @@ class ProductNormalizer implements NormalizerInterface
 
         $values = $values->filter(
             function ($value) use ($channelCode) {
-                return (!$value->getAttribute()->getScopable() || $value->getScope() == $channelCode);
+                return (!$value->getAttribute()->isScopable() || $value->getScope() == $channelCode);
             }
         );
 
@@ -167,7 +167,7 @@ class ProductNormalizer implements NormalizerInterface
 
         $values = $values->filter(
             function ($value) use ($localeCodes) {
-                return (!$value->getAttribute()->getTranslatable() || in_array($value->getLocale(), $localeCodes));
+                return (!$value->getAttribute()->isTranslatable() || in_array($value->getLocale(), $localeCodes));
             }
         );